修正一下cache逻辑
