+++ base/src/main/java/com/being/base/http/AsyncOkHttp.java
@@ -3,12 +3,11 @@ package com.being.base.http;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
-import android.util.SparseArray;
 
 import com.being.base.Constant;
 import com.being.base.http.callback.ResponseCallback;
 import com.being.base.http.exception.HttpRequestException;
-import com.being.base.http.intercept.CacheIntercept;
+import com.being.base.http.intercept.TryCacheInterceptor;
 import com.being.base.log.NHLog;
 import com.being.base.utils.DeviceInfoUtils;
 
@@ -74,11 +73,11 @@ public class AsyncOkHttp {
 
     private Handler mThreadHandler;
 
-    private CacheIntercept mCacheIntercept;
+    private TryCacheInterceptor mCacheIntercept;
 
     public AsyncOkHttp() {
         mThreadHandler = new Handler(Looper.getMainLooper());
-        mCacheIntercept = new CacheIntercept();
+        mCacheIntercept = new TryCacheInterceptor(mOkHttpClient);
         Dispatcher dispatcher = new Dispatcher();
         dispatcher.setMaxRequests(DEFAULT_MAX_CONNECTIONS);
         dispatcher.setMaxRequestsPerHost(DEFAULT_MAX_CONNECTIONS);
@@ -182,9 +181,7 @@ public class AsyncOkHttp {
             newCall.enqueue(new Callback() {
                 @Override
                 public void onFailure(final Call call, final IOException e) {
-                    if (!loadCache(call, responseCallback, callHandler)) {
-                        processFailure(call, e, responseCallback, callHandler);
-                    }
+                    processFailure(call, e, responseCallback, callHandler);
                 }
 
                 @Override
@@ -206,6 +203,7 @@ public class AsyncOkHttp {
 
     /**
      * 解析Response数据
+     *
      * @param call
      * @param response
      * @param responseCallback
@@ -260,6 +258,7 @@ public class AsyncOkHttp {
 
     /**
      * 网络请求成功
+     *
      * @param responseCallback
      * @param finalGsonType
      * @param response
@@ -284,6 +283,7 @@ public class AsyncOkHttp {
 
     /**
      * 网络请求失败
+     *
      * @param call
      * @param e
      * @param responseCallback
@@ -307,27 +307,27 @@ public class AsyncOkHttp {
         }
     }
 
-    private boolean loadCache(final Call call, final ResponseCallback responseCallback,
-                              CallHandler callHandler) {
-        if (call.request().method().equalsIgnoreCase("post")) {
-            return false;
-        }
-        boolean result = false;
-        //发生IO异常时,尝试从Http Cache中获取数据
-        Request request = call.request().newBuilder().cacheControl(CacheControl.FORCE_CACHE).build();
-        Response forceCacheResponse;
-        try {
-            forceCacheResponse = mOkHttpClient.newCall(request).execute();
-            if (forceCacheResponse.code() != 504) {
-                processResponse(call, forceCacheResponse, responseCallback, callHandler);
-                responseCallback.setCacheResponse(true);
-                result = true;
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return result;
-    }
+//    private boolean loadCache(final Call call, final ResponseCallback responseCallback,
+//                              CallHandler callHandler) {
+//        if (call.request().method().equalsIgnoreCase("post")) {
+//            return false;
+//        }
+//        boolean result = false;
+//        //发生IO异常时,尝试从Http Cache中获取数据
+//        Request request = call.request().newBuilder().cacheControl(CacheControl.FORCE_CACHE).build();
+//        Response forceCacheResponse;
+//        try {
+//            forceCacheResponse = mOkHttpClient.newCall(request).execute();
+//            if (forceCacheResponse.code() != 504) {
+//                processResponse(call, forceCacheResponse, responseCallback, callHandler);
+//                responseCallback.setCacheResponse(true);
+//                result = true;
+//            }
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+//        return result;
+//    }
 
     private void callExceptionFail(Response response, ResponseCallback responseCallback, Exception exception,
                                    CallHandler callHandler, Object jsonType) {
